Better Themes, Stats, and bugfixes (#1740)
* Fixed a bug where when clicking on a series rating for first time, the rating wasn't populating in the modal.

* Fixed a bug on Scroll mode with immersive mode, the bottom bar could clip with the book body.

* Cleanup some uses of var

* Refactored text as json into a type so I don't have to copy/paste everywhere

* Theme styles now override the defaults and theme owners no longer need to maintain all the variables themselves.

Themes can now override the color of the header on mobile devices via --theme-color and Kavita will now update both theme color as well as color scheme.

* Fixed a bug where last active on user stats wasn't for the particular user.

* Added a more accurate word count calculation and the ability to see the word counts year over year.

* Added a new table for long term statistics, like number of files over the years. No views are present for this data, I will add them later.

+++ API/DTOs/Statistics/PagesReadOnADayCount.cs
@@ -12,10 +12,9 @@ public class PagesReadOnADayCount<T> : ICount<T>
     /// <summary>
     /// Number of pages read
     /// </summary>
-    public int Count { get; set; }
+    public long Count { get; set; }
     /// <summary>
     /// Format of those files
     /// </summary>
     public MangaFormat Format { get; set; }
-
 }
